AP tag counter
Tag counter in the form of a small hovering box displayed in the bottom-left corner of the page.
When hidden, it shows a semi-transparent total tag count. Hover over it or any tags to reveal it fully.
Recommended tags (red line) shows a total number of them, while the number in brackets next to it shows approved recommendations just from you.
Requires an edit to line 19 and replacing the "insert_your_name" with your own display name.